Reorganize cohort layout into a single pane
* Rename requirements back to groups.
* Display groups in a single pane with criteria inline.
* Always show demographics while on the overview page and never on other
  pages within the cohort editor.
* Use text desriptions for filters with inline controls.
* Force undo/redo actions to navigate to the overview paeg rather than
  subpages (e.g. add criteria) so it's more clear what state the user is
  in.
* Don't ignore the excluded attribute of filters so their group counts
  better reflect the participants they actually match as part of the
  overall cohort.
