+++ ui/src/layout/gridBox.tsx
@@ -1,0 +1,6 @@
+import { styled } from "@mui/system";
+
+export const GridBox = styled("div")({
+  width: "100%",
+  height: "100%",
+});
